requirements.txt
django>=2.0
django-ckeditor
django-recaptcha
django-loginas
django-countries
django-extensions
gunicorn
django-webpack-loader
Pillow
django-imagekit
django-rest-framework
psycopg2-binary
django-crispy-forms
django-bootstrap-datepicker-plus
django_select2
django-colorful
django-js-reverse
